# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: My Health First API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medications:
    get:
      operationId: listMedications
      tags: [medications]
      summary: List all medications for current user
      responses:
        "200":
          description: List of medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [medications]
      summary: Add a new medication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicationBody"
      responses:
        "201":
          description: Medication created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
...
  /medications/{id}:
    get:
      operationId: getMedication
      tags: [medications]
      summary: Get a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "404":
          description: Not found
    patch:
      operationId: updateMedication
      tags: [medications]
      summary: Update a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicationBody"
      responses:
        "200":
          description: Updated medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "404":
          description: Not found
    delete:
      operationId: deleteMedication
      tags: [medications]
      summary: Delete a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /medications/{id}/taken:
    post:
      operationId: markMedicationTaken
      tags: [medications]
      summary: Mark a medication as taken for today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationReminder"
  /records:
    get:
      operationId: listRecords
      tags: [records]
      summary: List medical records
      responses:
        "200":
          description: List of medical records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalRecord"
    post:
      operationId: createRecord
      tags: [records]
      summary: Create a medical record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordBody"
      responses:
        "201":
          description: Record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /records/{id}:
    get:
      operationId: getRecord
      tags: [records]
      summary: Get a medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
        "404":
          description: Not found
    delete:
      operationId: deleteRecord
      tags: [records]
      summary: Delete a medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Medication:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        familyMemberId:
          type: ["integer", "null"]
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        times:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - dosage
        - frequency
        - times
        - startDate
        - isActive
        - createdAt
    CreateMedicationBody:
      type: object
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        times:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        familyMemberId:
          type: ["integer", "null"]
      required:
        - name
        - dosage
        - frequency
        - times
        - startDate
    UpdateMedicationBody:
      type: object
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        times:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
    MedicationReminder:
      type: object
      properties:
        id:
          type: integer
        medicationId:
          type: integer
        medicationName:
          type: string
        dosage:
          type: string
        scheduledTime:
          type: string
        taken:
          type: boolean
        takenAt:
          type: ["string", "null"]
        date:
          type: string
        familyMemberName:
          type: ["string", "null"]
      required:
        - id
        - medicationId
        - medicationName
        - dosage
        - scheduledTime
        - taken
        - date
    MedicalRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        familyMemberId:
          type: ["integer", "null"]
        title:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        recordDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - title
        - category
        - recordDate
        - createdAt
    CreateRecordBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        recordDate:
          type: string
        familyMemberId:
          type: ["integer", "null"]
      required:
        - title
        - category
        - recordDate
...